Compute orientations of physical objects (board, laser, mirrors) (#7)
* find location of board based on laser samples

* Compute locations and orientations of physical objects (the Go board, the laser pointer, the mirrors)

* Fix some warnings under gcc, use -O2 in tests because eigen is slooowww

---------

Authored-by: jackson stogel <jtstogel@jtstogel.jacksonstogel>

+++ src/util/union_find_test.cc
@@ -18,7 +18,7 @@ TEST(UnionFind, Union) {
 
 TEST(UnionFind, TransitiveSetMembership) {
   UnionFind uf(1000);
-  for (int i = 0; i < uf.Size() - 1; i++) {
+  for (size_t i = 0; i < uf.Size() - 1; i++) {
     uf.Union(i, i + 1);
   }
   EXPECT_EQ(uf.FindSet(0), uf.FindSet(999));
